requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
numpy>=1.24.0
tiktoken>=0.5.0
openai>=1.0.0
python-dotenv>=1.0.0
lxml>=4.9.0
tqdm>=4.65.0
fake-useragent>=1.4.0
